backend/requirements.txt
fastapi[all]
uvicorn[standard]
python-multipart
pyyaml
protobuf<5,>=4.25.3
grpcio 
grpcio-tools
mypy-protobuf
psycopg2-binary
timescaledb
torch
onnx
onnxruntime

numpy

opencv-python
mediapipe
openai-whisper
moviepy
tqdm

langchain 
langchain-openai
langchain-community
python-dotenv

